feat(policy): Refactor policy engine and message bus
This commit addresses several comments from PR #7835.

- Refactors `removeRulesForTool` in `PolicyEngine` to use `.filter()` for better readability.
- Extracts `stableStringify` into its own utility file and makes `ruleMatches` a file-level function.
- Adds structural types to message objects in `message-bus.test.ts` to improve type safety.
- Makes `ToolExecutionSuccess` and `ToolExecutionFailure` interfaces generic for more flexible typing.
- Fixes failing tests in `message-bus.test.ts`.
